requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
scikit-learn
pydantic